add canPrint authorization on print slips route
