+++ app/(no-layout)/slips/[id]/ClientMultipagePdf.tsx
@@ -1,0 +1,15 @@
+'use client';
+import { useParams } from 'next/navigation';
+
+export default function MultiPagePdf() {
+  const { id } = useParams();
+  return (
+    <div style={{ height: '100vh' }}>
+      <iframe
+        src={`${process.env.NEXT_PUBLIC_APP_BASEPATH ?? ''}/api/slipsPdf/${id}`}
+        title="Printing Slips"
+        style={{ width: '100%', height: '100%', border: 'none' }}
+      />
+    </div>
+  );
+}
